==> step1_audio_transcription.ipynb <==
PROJECT_HOME = "."

# # For Colab

# PROJECT_HOME = "/content/drive/My Drive/Projects/LLM-MCI-detection"

# # Google Drive storage setup
# from google.colab import drive
# drive.mount('/content/drive', force_remount=True)

%%capture
%pip install git+https://github.com/openai/whisper.git

!sudo apt update >/dev/null 2>&1 && sudo apt install ffmpeg >/dev/null 2>&1

import os
import whisper
from tqdm import tqdm

output_dir = os.path.join(PROJECT_HOME, "transcripts")
os.makedirs(output_dir, exist_ok=True)

audio_file_paths = []

for folder_name in ["dementia", "control"]:

    data_dir = os.path.join(PROJECT_HOME, "Pitt", "cookie", folder_name)
    for file_name in os.listdir(data_dir):
        if file_name.endswith('.mp3'):
            file_path = os.path.join(data_dir, file_name)
            audio_file_paths.append(file_path)

assert len(audio_file_paths) == len(set(audio_file_paths))

model = whisper.load_model("large-v3", download_root=os.path.join(PROJECT_HOME, "whisper"))

subject_records = []
for file_path in tqdm(audio_file_paths, "Generating audio transcript"):
    result = model.transcribe(file_path, prompt="Umm, let me think like, hmm... Okay, here's what I'm, like, thinking.")
    text = result["text"].strip()
    with open(os.path.join(output_dir, os.path.basename(file_path).replace(".mp3", ".txt")), "w") as transcript_f:
        transcript_f.write("%s\n" % text)
